pyproject.toml
[project]
name = "ai-agent-tool-server"
version = "0.1.0"
description = ""
authors = [
    {name = "João Morossini",email = "joaomorossini@duck.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.13",
    "langchain-openai>=0.0.2",
    "langgraph>=0.0.19",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",  # For PostgreSQL support
    "python-multipart>=0.0.6",
    "typing-extensions>=4.9.0",
    "asyncio>=3.4.3",
    "python-jose[cryptography]>=3.3.0",
    "loguru>=0.7.2",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "httpx>=0.26.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
